Improve instructions for building GraalVM native image
Improve instructions for building GraalVM native image: 
Building in a docker container relies on the artifacts being available before you start the docker container. Added a description of that. 
native image can use more memory than the docker container is configured with, and then it fails with the error code 137 -- Add a comment about that with a link to the help on configuring the memory for Docker.